requirements.txt
# TFG Project Dependencies
# Minimal subset of trading project requirements for local academic work.
# Excludes: google-cloud-*, kubernetes, prometheus, telegram, websockets,
#           httpx, aiohttp, fastapi, uvicorn, and other remote infrastructure deps.

# Core Data & Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Reinforcement Learning
stable-baselines3>=2.3.0
gymnasium>=0.29.0
shimmy>=1.3.0
torch>=2.0.0
tensorboard>=2.14.0
tqdm>=4.65.0
rich>=13.0.0

# Hyperparameter Optimization
optuna>=3.5.0

# ML Tracking
mlflow>=2.9.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Jupyter Notebooks
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
psutil>=5.9.0
